Fixes colorbar bug in contour, other improvements
Changes how colorbars are handled in kw_contour, a little more
flexible as far as setting discrete colorbar marks with a large
number of colors for contouring.

Improve position of mini_map on plot using x/y char sizes.

Changes logic in nearest neighbor interpolation as it was not
acting as I thought it should; perhaps I broke it for old code?

v0.0.2