Improves CT loading, grid lines, text
The KWLOADCT procedure has been added, which is a wrapper for
LOADCT procedure that implements the BOTTOM keyword in the way
I think it should work. The BOTTOM keyword is LOADCT does not seem
to start loading colors at the specified point in the color table.
The new KWLOADCT actually starts loading colors at requested index.

The KW_MAP_GRID procedure has also been updated to fixes issues
where grid lines were not draw because the end points for the grid
lines were outside of the map limits.

The KW_TEXT procedure has been updated to improve the drawing of
the background behind text; the ALIGNMENT keyword had not been
implemented properly so the background would not be placed in the
correct spot.

V0.3.2